requirements.txt
PyQt5>=5.15.0
pyttsx3>=2.90
pygame>=2.0.0
transformers>=4.18.0
torch>=1.10.0
duckduckgo_search>=2.0.0
requests>=2.25.0
beautifulsoup4>=4.9.0
sentence-transformers>=2.2.0
numpy>=1.19.0
tqdm>=4.50.0
pillow>=8.0.0
pyinstaller>=5.0.0